+++ FirebaseAdmin/FirebaseAdmin.Snippets/FirebaseAuthSnippets.cs
@@ -63,7 +63,7 @@ namespace FirebaseAdmin.Snippets
             // [START get_user_by_id]
             UserRecord userRecord = await FirebaseAuth.DefaultInstance.GetUserAsync(uid);
             // See the UserRecord reference doc for the contents of userRecord.
-            Console.WriteLine("Successfully fetched user data: {0}", userRecord.Uid);
+            Console.WriteLine($"Successfully fetched user data: {userRecord.Uid}");
             // [END get_user_by_id]
         }
 
@@ -72,7 +72,7 @@ namespace FirebaseAdmin.Snippets
             // [START get_user_by_email]
             UserRecord userRecord = await FirebaseAuth.DefaultInstance.GetUserByEmailAsync(email);
             // See the UserRecord reference doc for the contents of userRecord.
-            Console.WriteLine("Successfully fetched user data: {0}", userRecord.Uid);
+            Console.WriteLine($"Successfully fetched user data: {userRecord.Uid}");
             // [END get_user_by_email]
         }
 
@@ -81,8 +81,62 @@ namespace FirebaseAdmin.Snippets
             // [START get_user_by_phone]
             UserRecord userRecord = await FirebaseAuth.DefaultInstance.GetUserByPhoneNumberAsync(phoneNumber);
             // See the UserRecord reference doc for the contents of userRecord.
-            Console.WriteLine("Successfully fetched user data: {0}", userRecord.Uid);
+            Console.WriteLine($"Successfully fetched user data: {userRecord.Uid}");
             // [END get_user_by_phone]
+        }
+
+        internal static async Task CreateUserAsync()
+        {
+            // [START create_user]
+            UserRecordArgs args = new UserRecordArgs()
+            {
+                Email = "user@example.com",
+                EmailVerified = false,
+                PhoneNumber = "+11234567890",
+                Password = "secretPassword",
+                DisplayName = "John Doe",
+                PhotoUrl = "http://www.example.com/12345678/photo.png",
+                Disabled = false,
+            };
+            UserRecord userRecord = await FirebaseAuth.DefaultInstance.CreateUserAsync(args);
+            // See the UserRecord reference doc for the contents of userRecord.
+            Console.WriteLine($"Successfully created new user: {userRecord.Uid}");
+            // [END create_user]
+        }
+
+        internal static async Task CreateUserWithUidAsync()
+        {
+            // [START create_user_with_uid]
+            UserRecordArgs args = new UserRecordArgs()
+            {
+                Uid = "some-uid",
+                Email = "user@example.com",
+                PhoneNumber = "+11234567890",
+            };
+            UserRecord userRecord = await FirebaseAuth.DefaultInstance.CreateUserAsync(args);
+            // See the UserRecord reference doc for the contents of userRecord.
+            Console.WriteLine($"Successfully created new user: {userRecord.Uid}");
+            // [END create_user_with_uid]
+        }
+
+        internal static async Task UpdateUserAsync(string uid)
+        {
+            // [START update_user]
+            UserRecordArgs args = new UserRecordArgs()
+            {
+                Uid = uid,
+                Email = "modifiedUser@example.com",
+                PhoneNumber = "+11234567890",
+                EmailVerified = true,
+                Password = "newPassword",
+                DisplayName = "Jane Doe",
+                PhotoUrl = "http://www.example.com/12345678/photo.png",
+                Disabled = true,
+            };
+            UserRecord userRecord = await FirebaseAuth.DefaultInstance.UpdateUserAsync(args);
+            // See the UserRecord reference doc for the contents of userRecord.
+            Console.WriteLine($"Successfully updated user: {userRecord.Uid}");
+            // [END update_user]
         }
 
         internal static async Task DeleteUserAsync(string uid)
@@ -91,6 +145,33 @@ namespace FirebaseAdmin.Snippets
             await FirebaseAuth.DefaultInstance.DeleteUserAsync(uid);
             Console.WriteLine("Successfully deleted user.");
             // [END delete_user]
+        }
+
+        internal static async Task ListAllUsersAsync()
+        {
+            // [START list_all_users]
+            // Start listing users from the beginning, 1000 at a time.
+            var pagedEnumerable = FirebaseAuth.DefaultInstance.ListUsersAsync(null);
+            var responses = pagedEnumerable.AsRawResponses().GetEnumerator();
+            while (await responses.MoveNext())
+            {
+                ExportedUserRecords response = responses.Current;
+                foreach (ExportedUserRecord user in response.Users)
+                {
+                    Console.WriteLine($"User: {user.Uid}");
+                }
+            }
+
+            // Iterate through all users. This will still retrieve users in batches,
+            // buffering no more than 1000 users in memory at a time.
+            var enumerator = FirebaseAuth.DefaultInstance.ListUsersAsync(null).GetEnumerator();
+            while (await enumerator.MoveNext())
+            {
+                ExportedUserRecord user = enumerator.Current;
+                Console.WriteLine($"User: {user.Uid}");
+            }
+
+            // [END list_all_users]
         }
 
         internal static async Task SetCustomUserClaimsAsync(string uid)
@@ -126,6 +207,43 @@ namespace FirebaseAdmin.Snippets
             UserRecord user = await FirebaseAuth.DefaultInstance.GetUserAsync(uid);
             Console.WriteLine(user.CustomClaims["admin"]);
             // [END read_custom_user_claims]
+        }
+
+        internal static async Task SetCustomUserClaimsScriptAsync()
+        {
+            // [START set_custom_user_claims_script]
+            UserRecord user = await FirebaseAuth.DefaultInstance
+                .GetUserByEmailAsync("user@admin.example.com");
+            // Confirm user is verified.
+            if (user.EmailVerified)
+            {
+                var claims = new Dictionary<string, object>()
+                {
+                    { "admin", true },
+                };
+                await FirebaseAuth.DefaultInstance.SetCustomUserClaimsAsync(user.Uid, claims);
+            }
+
+            // [END set_custom_user_claims_script]
+        }
+
+        internal static async Task SetCustomUserClaimsIncrementalAsync()
+        {
+            // [START set_custom_user_claims_incremental]
+            UserRecord user = await FirebaseAuth.DefaultInstance
+                .GetUserByEmailAsync("user@admin.example.com");
+            // Add incremental custom claims without overwriting the existing claims.
+            object isAdmin;
+            if (user.CustomClaims.TryGetValue("admin", out isAdmin) && (bool)isAdmin)
+            {
+                var claims = new Dictionary<string, object>(user.CustomClaims);
+                // Add level.
+                claims["level"] = 10;
+                // Add custom claims for additional privileges.
+                await FirebaseAuth.DefaultInstance.SetCustomUserClaimsAsync(user.Uid, claims);
+            }
+
+            // [END set_custom_user_claims_incremental]
         }
     }
 }